% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.variables.R
\name{analyze.p}
\alias{analyze.p}
\title{A function which will execute pwr.sample function if the function is missing the power variable.}
\usage{
analyze.p(sdA0, sdA1, sdB0, sdB1, r1, alpha, m1, m2, power, n, ind)
}
\arguments{
\item{sdA0}{A number.}

\item{sdA1}{A number.}

\item{sdB0}{A number.}

\item{sdB1}{A number.}

\item{r1}{A number.}

\item{alpha}{A number.}

\item{m1}{A number.}

\item{m2}{A number.}

\item{power}{A number.}

\item{n}{A number.}

\item{ind}{A quoted character string ("sample" OR "diff").}
}
\description{
This function takes all standard inputs as well as a given indicator variable "sample" or "diff".  If there is no input for power, ind.power will be executed
}
\examples{
analyze.p(sdA0= 2, sdA1= 2, sdB0= 2, sdB1= 2, r1= c(.6,.7,.8), alpha= .05, m1= .54, m2= 55, n= c(60,64,68), ind = "sample")
}
