% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.clean.R
\name{clean.1}
\alias{clean.1}
\title{Cleans the list which is the output of analyze.input}
\usage{
clean.1(obj)
}
\arguments{
\item{obj}{A list.}
}
\value{
A list of either length 1 or 2.
}
\description{
This function takes a three element list with two NULL elements and will either eliminate the first element if it is NULL or eliminate the last two if it is not.
}
\examples{
clean.1(obj)
}
