% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finding Variables.R
\name{diff.find}
\alias{diff.find}
\title{Find the difference of means based on given inputs.}
\usage{
\method{diff}{find}(sdA0, sdA1, sdB0, sdB1, r1, alpha, power, N)
}
\arguments{
\item{sdA0}{A number.}

\item{sdA1}{A number.}

\item{sdB0}{A number.}

\item{sdB1}{A number.}

\item{r1}{A number.}

\item{alpha}{A number.}

\item{power}{A number.}

\item{n}{A number.}
}
\value{
sammple size of the model
}
\description{
Find the difference of means based on given inputs.
}
\examples{
diff.find(2,2,2,2,.5,.05,.8,64)
}
