% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indicator.sorting.R
\name{ind.sample}
\alias{ind.sample}
\title{A function which executes one of the plot.sampfind functions based on a given indicator variable.}
\usage{
ind.sample(sdA0, sdA1, sdB0, sdB1, r1, alpha, m1, m2, power, n, ind)
}
\arguments{
\item{sdA0}{A number.}

\item{sdA1}{A number.}

\item{sdB0}{A number.}

\item{sdB1}{A number.}

\item{r1}{A number.}

\item{alpha}{A number.}

\item{m1}{A number.}

\item{m2}{A number.}

\item{power}{A number.}

\item{ind}{A quoted character string ("power" OR "diff").}
}
\description{
This function takes standard inputs as well as a given indicator variable "power" or "diff" in order to distinguish and execute the correct function.
}
\examples{
ind.sample(sdA0= 2, sdA1= 2, sdB0= 2, sdB1= 2, r1= c(.6,.7,.8), alpha= .05, m1= .54, m2= 55, power=  c(.6,.7,.8), ind = "power")
}
