% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solving.Plotting.R
\name{plot.diffind.n}
\alias{plot.diffind.n}
\title{Creates a data frame and plot of grouped observations for missing mean differences based on sample size.}
\usage{
\method{plot}{diffind.n}(sdA0, sdA1, sdB0, sdB1, r1, alpha, power, n)
}
\arguments{
\item{sdA0}{A number.}

\item{sdA1}{A number.}

\item{sdB0}{A number.}

\item{sdB1}{A number.}

\item{r1}{A number.}

\item{alpha}{A number.}

\item{power}{A number.}

\item{n}{A number.}
}
\description{
This function takes numbers and vectors, turns them into a data frame and solves for every combination of input.  It also plots output by sample size and groups observations by similar non sample-inputs.
}
\examples{
plot.diffind.n(2,2,2,2,.5,.05,.8,c(60,64,68))
}
