% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solving.Plotting.R
\name{plot.diffind.p}
\alias{plot.diffind.p}
\title{Creates a data frame and plot of grouped observations for missing mean differences based on power.}
\usage{
\method{plot}{diffind.p}(sdA0, sdA1, sdB0, sdB1, r1, alpha, power, n)
}
\arguments{
\item{sdA0}{A number.}

\item{sdA1}{A number.}

\item{sdB0}{A number.}

\item{sdB1}{A number.}

\item{r1}{A number.}

\item{alpha}{A number.}

\item{power}{A number.}

\item{n}{A number.}
}
\description{
This function takes numbers and vectors, turns them into a data frame and solves for every combination of input.  It also plots output by power and groups observations by similar non power-inputs.
}
\examples{
plot.diffind.p(2,2,2,2,.5,.05,c(.6,.7,.8),64,)
}
