% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solving.Plotting.R
\name{plot.pwrfind.n}
\alias{plot.pwrfind.n}
\title{Creates a data frame and plot of grouped observations for missing powers based on difference of sample size.}
\usage{
\method{plot}{pwrfind.n}(sdA0, sdA1, sdB0, sdB1, r1, alpha, m1, m2, n)
}
\arguments{
\item{sdA0}{A number.}

\item{sdA1}{A number.}

\item{sdB0}{A number.}

\item{sdB1}{A number.}

\item{r1}{A number.}

\item{alpha}{A number.}

\item{m1}{A number.}

\item{m2}{A number.}

\item{n}{A number.}
}
\description{
This function takes numbers and vectors, turns them into a data frame and solves for every combination of input.  It also plots output by sample size and groups observations by similar non sample-inputs.
}
\examples{
plot.pwrfind.n(2,2,2,2,.5,.05,54,55,c(60,64,68))
}
