% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finding Variables.R
\name{power.find}
\alias{power.find}
\title{Find the power based on given inputs.}
\usage{
power.find(sdA0, sdA1, sdB0, sdB1, r1, alpha, m1, m2, n)
}
\arguments{
\item{sdA0}{A number.}

\item{sdA1}{A number.}

\item{sdB0}{A number.}

\item{sdB1}{A number.}

\item{r1}{A number.}

\item{alpha}{A number.}

\item{m1}{A number.}

\item{m2}{A number.}

\item{n}{A number.}
}
\value{
sammple size of the model
}
\description{
Find the power based on given inputs.
}
\examples{
power.find(2,2,2,2,.5,.05,54,55,64)
}
