% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finding Variables.R
\name{sample.find}
\alias{sample.find}
\title{Find the sample size based on given inputs.}
\usage{
sample.find(sdA0, sdA1, sdB0, sdB1, r1, alpha, m1, m2, power)
}
\arguments{
\item{sdA0}{A number.}

\item{sdA1}{A number.}

\item{sdB0}{A number.}

\item{sdB1}{A number.}

\item{r1}{A number.}

\item{alpha}{A number.}

\item{m1}{A number.}

\item{m2}{A number.}

\item{power}{A number}
}
\value{
sammple size of the model
}
\description{
Find the sample size based on given inputs.
}
\examples{
sample.find(2,2,2,2,.5,.05,55,54,.8)
}
