% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seperate.solve.R
\name{sep_solve_dp}
\alias{sep_solve_dp}
\title{Seperates data frame row and finds  mean difference based on power}
\usage{
sep_solve_dp(dfl)
}
\arguments{
\item{dfl}{A data frame row.}
}
\value{
a vector of solved mean differences.
}
\description{
This function is meant to allow data frames to be solved by row with apply.  It transforms a row into a matrix and uses sample.find across the columns of this one row matrix.
}
\examples{
sep_solve_dp(df[1,])
}
