% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seperate.solve.R
\name{sep_solve_nd}
\alias{sep_solve_nd}
\title{Seperates data frame row and finds sample based on mean difference}
\usage{
sep_solve_nd(dfl)
}
\arguments{
\item{dfl}{A data frame row.}
}
\value{
a vector of solved sample sizes.
}
\description{
This function is meant to allow data frames to be solved by row with apply.  It transforms a row into a matrix and uses sample.find across the columns of this one row matrix.
}
\examples{
sep_solve_nd(df[1,])
}
